backend/requirements.txt
asgiref==3.11.1
celery[redis]==5.4.0
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
pyotp==2.9.0
python-dateutil==2.9.0
qrcode[pil]==8.0
sentry-sdk[django]==2.22.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.8.2
